Evaluator V1 + Input Validation
Adds evaluator layout and scripts for input validation that can be used as a template for all games.

Conditional_DisplayText_OnClick can have any button, error text, and can be reconfigured for any gameobject script that returns a bool to run plug 'n' play style. The idea is you use this on every game and make a custom script, like Validation_Evaluator, that plugs into it and just has the parts unique to that game in a separate script.

It's a bit small now but I figure it's a good foundation to modularize these class functions separately now before things start getting more complex later.