Data Manager
Big update!

Creates the Data Manager. This class uses static variables to preserve data between scenes. It has not-so-static functions that can be called to load and save to it.

It might make since to divide this into a user manager and game data manager later on once we have a clearer picture of how the backend will work.

This current implementation preserves teacher, assessor, child, scores, and response and displays it all on the Grader screen.

There is a key string called 'currentScene' that controls the logic flow for the class. This relies on other classes to update it right before they load the next scene - Unity technically has Scene transition events but they're a bit wonky. Trading a dependency to do this job quicker and cleaner.

Also made the UserInfo text black for readability.